gate/requirements.txt
socketIO-client==0.5.6
requests
websocket_client
